+++ cardigan_beta/stories/components/Links.js
@@ -1,9 +1,9 @@
 import { storiesOf } from '@storybook/react';
 
-import PrimaryLink from '../../common/views/components/Links/PrimaryLink/PrimaryLink';
-import SecondaryLink from '../../common/views/components/Links/SecondaryLink/SecondaryLink';
+import PrimaryLink from '../../../common/views/components/Links/PrimaryLink/PrimaryLink';
+import SecondaryLink from '../../../common/views/components/Links/SecondaryLink/SecondaryLink';
 
-const links = storiesOf('Links', module);
+const links = storiesOf('Components', module);
 links.add('Primary link', () => (
   <PrimaryLink url={`#`} name={`Take me to the moon`} />
 ));
